+++ src/components/commons/TemperatureBar/TemperatureBar.style.ts
@@ -33,35 +33,6 @@ export const fireIconStyle = css`
     width: 1.6rem;
     height: 1.6rem;
 `
-export const popoverContainer = css`
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    position: relative;
-    bottom: 1.75rem;
-`
-export const popoverContentStyle = css`
-    display: flex;
-    align-items: center;
-    width: 23.1rem;
-    height: 6.7rem;
-    padding: 0.8rem 1.6rem;
-    border-radius: 12px;
-    background-color: ${theme.color.object.hero};
-    color: ${theme.color.base.white};
-    ${theme.font.caption1};
-    word-wrap: break-word;
-`
-export const popoverPinStyle = css`
-    height: 1.2rem;
-    color: ${theme.color.object.hero};
-    margin-top: -0.3rem;
-`
-export const temperatureInformationIconStyle = css`
-    width: 1.7rem;
-    height: 1.7rem;
-    cursor: pointer;
-`
 export const temperatureBarContainer = css`
   display: flex;
   width: 37.9rem;
